+++ pirel/_guess.py
@@ -4,6 +4,7 @@ import abc
 import csv
 import datetime
 import inspect
+import logging
 import random
 from typing import Callable, Iterable, Optional, TextIO
 
@@ -12,11 +13,15 @@ from rich.console import Console
 from rich.prompt import DefaultType, Prompt
 from rich.text import Text
 
+import pirel
+
 from .releases import PythonRelease
 
 STATS_DIR = platformdirs.user_data_path("pirel")
 STATS_FILENAME = "guess_stats.csv"
 STATS_FIELDNAMES = ["time", "question_cls", "target_release", "score"]
+
+logger = logging.getLogger("pirel")
 
 
 class PirelPrompt(Prompt):
@@ -130,12 +135,14 @@ class Question(abc.ABC):
     target_release: PythonRelease
     get_target_field: Callable[[PythonRelease], str]
 
-    def __init__(self, releases: list[PythonRelease], console: Console):
+    def __init__(self, releases: list[PythonRelease]):
         if not hasattr(self, "target_release"):
             self.target_release = random.choice(releases)
         self.releases = releases
-        self.console = console
         self.choices = self.build_choices()
+
+    def __repr__(self) -> str:
+        return f"{self.__class__.__name__}({self.target_release.version})"
 
     @property
     def correct_answer(self) -> str:
@@ -212,15 +219,18 @@ class Question(abc.ABC):
         prompt = PirelPrompt(
             self.format_question(),
             choices=self.choices,
-            console=self.console,
+            console=pirel.CONTEXT.rich_console,
         )
         answer = prompt(stream=PirelPrompt.stream)
         if answer == self.correct_answer:
-            self.console.print(f"[prompt.choices]{answer}[/] is [green]correct![/]")
+            pirel.CONTEXT.rich_console.print(
+                f"[prompt.choices]{answer}[/] is [green]correct![/]"
+            )
             return True
         else:
-            self.console.print(
-                f"[prompt.choices]{answer}[/] is [red]wrong![/] (Correct answer: [prompt.choices]{self.correct_answer}[/])"
+            pirel.CONTEXT.rich_console.print(
+                f"[prompt.choices]{answer}[/] is [red]wrong![/]"
+                f" (Correct answer: [prompt.choices]{self.correct_answer}[/])"
             )
             return False
 
@@ -229,10 +239,10 @@ class LatestVersionQuestion(Question):
     question: str = "What is the latest stable version of Python?"
     __doc__ = question
 
-    def __init__(self, releases: list[PythonRelease], console: Console):
+    def __init__(self, releases: list[PythonRelease]):
         self.target_release = max(filter(lambda x: x._status == "bugfix", releases))
         self.get_target_field = lambda x: x.version
-        super().__init__(releases, console)
+        super().__init__(releases)
 
     def incorrect_choices(self) -> list[PythonRelease]:
         return self.generate_incorrect_choices(lambda x: x.version.startswith("3"))
@@ -242,10 +252,10 @@ class VersionDateQuestion(Question):
     question: str = "When was Python {version} released?"
     __doc__ = question
 
-    def __init__(self, releases: list[PythonRelease], console: Console):
+    def __init__(self, releases: list[PythonRelease]):
         self.target_release = random.choice(releases)
         self.get_target_field = lambda x: x._released.isoformat()
-        super().__init__(releases, console)
+        super().__init__(releases)
 
     def format_question(self) -> str:
         return self.question.format(version=self.target_release.version)
@@ -255,10 +265,10 @@ class DateVersionQuestion(Question):
     question: str = "Which version of Python was released on {release_date}?"
     __doc__ = question
 
-    def __init__(self, releases: list[PythonRelease], console: Console):
+    def __init__(self, releases: list[PythonRelease]):
         self.target_release = random.choice(releases)
         self.get_target_field = lambda x: x.version
-        super().__init__(releases, console)
+        super().__init__(releases)
 
     def format_question(self) -> str:
         return self.question.format(release_date=self.target_release._released)
@@ -268,9 +278,9 @@ class ReleaseManagerVersionQuestion(Question):
     question: str = "Who was the release manager for Python {version}?"
     __doc__ = question
 
-    def __init__(self, releases: list[PythonRelease], console: Console):
+    def __init__(self, releases: list[PythonRelease]):
         self.get_target_field = lambda x: x._release_manager
-        super().__init__(releases, console)
+        super().__init__(releases)
 
     def format_question(self) -> str:
         return self.question.format(version=self.target_release._version)
@@ -285,6 +295,13 @@ def store_question_score(question: Question, score: int) -> None:
         STATS_DIR.mkdir(parents=True)
 
     stats_file = STATS_DIR / STATS_FILENAME
+    question_type = question.__class__.__name__
+    logger.info(
+        "Storing score %d for question type %s in file '%s'",
+        score,
+        question_type,
+        stats_file,
+    )
     with stats_file.open("a", newline="") as file:
         writer = csv.DictWriter(
             file, fieldnames=STATS_FIELDNAMES, quoting=csv.QUOTE_NONNUMERIC
@@ -301,12 +318,8 @@ def store_question_score(question: Question, score: int) -> None:
         )
 
 
-def get_random_question(releases: list[PythonRelease], console: Console) -> Question:
+def get_random_question() -> Question:
     """Randomly picks one of the available questions.
-
-    Args:
-        releases (list[PythonRelease]): A list of Python release objects.
-        console (Console): A rich Console instance.
 
     Returns:
         type[Question]: A class of type `Question`.
@@ -318,17 +331,4 @@ def get_random_question(releases: list[PythonRelease], console: Console) -> Ques
             if inspect.isclass(var) and issubclass(var, Question) and var != Question
         ]
     )
-    return question_cls(releases, console)
-
-
-def ask_random_question(releases: list[PythonRelease], console: Console) -> None:
-    """Prompt the user with one of the available questions and store the answer in user
-    data.
-
-    Args:
-        releases (list[PythonRelease]): A list of Python release objects.
-        console (Console): A rich Console instance.
-    """
-    question = get_random_question(releases, console)
-    score = int(question.ask())
-    store_question_score(question, score)
+    return question_cls(pirel.CONTEXT.releases.to_list())